Error: asio::any_io_executor::any:io_executor: No overload function could convert arguments
This error is displayed when building the project with MSVC on Windows.
The program builds on Ubuntu wsl.
Building the programm in windows-x64-static and windows-x64 gives the
same output. Building it with Md instead of Mt also.
I tried to force the use of Boost asio instead of asio, same error.
It seems that the path handeling of windows and linux when asio and
boost-asio are both included in a project is different.