First Version of the WS
This is the first version of the websocket.
Choosing UUIDs and starting the data aquisation
is handle completely from the CLI Tool.
Data is send in JSON format with 1000 samples each
package. As the data aquisation starts without a WS
connection and is popped from a deque, the send data
starts with the first value that was measured when the CLI
Tool started the WS.