API Endpoints / Sampling:
Changes:

The UUIDs can printed for the Client via the /UUID
endpoint of the Crow API on Port 8080.
The help message on how to use the websocket can
be printed via the /help endpoint of the API.

TEMP: Sampling Rate can be set as the last option when starting
a websocket with UUIDs. The sampling rate feature is not fully
developed.