Corrected JSON formats of Websocket and REST
Problem: The current formats for sending data were not as in the
API Documentation and not suitable for the frontend

Solution: The data format send from the UUID REST API endpoint was fixed
to fit the documented JSON format. The json format of data send via the
websocket was changed to include the devices UUIDs.