TEMP: Refactoring : Measurement Object
Setting up a Measurement object that includes all necessary information
about a measurement. This includes:
UUIDs of the Scopes, format, data destination, sampling rate,optinal
filepath
This is used to have a better workflow when working with different
formats and destination as well as samplerates. No new features are
currently included.

Breaking changes in this commit: For CLI measurements the sampling rate
is set to 10 Sa/s instead of 100000 Sa/s.