BREAKING CHANGES + Refactoring:
Breaking Change: It is no longer possible to set the UUIDs for the WS
from the CLI tool, they have to be set via the client.
Reason: Setting the UUIDs from the CLI Tool was a necessary step for the
development of the BE but is no longer used.

Refactoring:

Setting up the "measurement" object. This is further developed by using
the measurement object in all functions needed instead of using global
variables.