CSV Format for Websocket
Changes:

The dataformat that is send via the websocket can now be set via the
client.

Formats that are available are json and csv. With the csv format the
samples are send sample by sample. The csv format does not include
informations about the device UUID.
The default format is JSON.