Fix: Reset Queues/Format Change + WS Binary Format
Bug: When opening a WS the first connection to the WS works fine. When
connecting a second time the Queues where not reset. Also due to the
static Writer boolean a Writer with a different format was never
started. This is fixed now.

Changes: Reset of all Queues when closing a WS connection. Writer
restarts with a new connection. Binary format for the WS is available
now.

Current Problem: Sometimes when switching from json to binary or from
csv to binary the programm crashes unexpected.