Fix: SW crash when switching formats
Problem: When switching to a different format on the client side and
connecting again to the websocket, the sw sometimes crashed.

Solution: Due to wrong state management the writer function never
stopped and had an access error to the queue. This was fixed by changing
the state in which the function works.
Fix: SW crash when switching formats

Problem: When switching to a different format on the client side and
connecting again to the websocket, the sw sometimes crashed.

Solution: Due to wrong state management the writer function never
stopped and had an access error to the queue. This was fixed by changing
the state in which the function works.