Feature: Receive additional information via get_info endpoint (#34)
* Feature: Receive additional information via API

Problem: For a possible client it is not possible to get firmware, hardware and
calibration information via an API.

Changes: endpoint: v1/get_info

This endpoint provides the hardware, firmware and calibration
information, as well as the UUIDs and colors as json.

* Doc: README.md v1/get_info enpoint

Updated the README.md to include the new get_info endpoint .

Also updated the link to the API description as the repo name changed.

* Fix: Port and formatter

The code currently did not have the correct format. Corrected with
astyle.

Added the - to the command -w -p in the README.md