doc(update): Comment code (#43)
* doc(update): Comment code

Comment code with code standard comments for functions for better understanding and readability

This is step 1 of a refactoring to make the code less buggy and more maintainable

* doc(comment): Document how copyOut works

It was not clear if copyOut deletes all samples that are copied from the
sampler. This is important as it makes sure no samples are copied twice
in captureData. As the submodule is closed source this behavior should
be documented to understand the function

* fix(output): Delete "Format is json"

Delete "Format is json" console.log output.

The output was not necessary.