Fix: Device search (#44)
* fix(bug): Delete stop in searchDevices

Delete stop time of 2 seconds in searchDevices after devices were found.
This should not be part of searchDevices and if needed with this
function should be implemented outside.

* fix(bug): Delete Search Devices after every batch

Delete SearchDevices() and SelectDevices() function in while loop of
start function.

Reason:
When PrintOrWrite data is done with a batch the SearchDevices function
was called again. This resets all devices for a moment and will lead to
data loss while a measurement is taken.