fix(bug): Add writer stop on ws disconnect (#45)
Add writer stop function in disconnect websocket function. Add
ControlWriter class to control the writer object.

Problem:
The Writer was not destroyed on ws disconnection which could lead to
data that is still stored in the writer when opening a new connection.
Destroying the writer on ws disconnection should fix this.