Enhancement: Start Sending via jthread
Add logic to start data sending via WS via a jthread instead of a
thread.
Add WSContext that holds the jthread and measurement.

Instead of booleans for statehandling and joining the threads in
different functions, jthreads can be closed via a stop_request -->
State and thread handling gets more maintainable