feat(save): Add save option to WS
Add option to save data after a measurement in a file .
Add API endpoint to download file after saving

Note: This is a trivial implementation not using a ring buffer system for faster developing
with the current architecture of the backend. Savings cant be made while
a measurement is running. If the RAM is
full no trigger is set that will stop the measurement. The architecture
is chosen to implement the functionality for OmnAIView for the time
before NaDI is finished.
It is expected to be a work around for a few weeks. If this
implementation should be further improved this should be refactored.