init: Doc for saved file formats (#54)
* init): Doc for saved file formats

Add doc for saved file formats in docs\FileFormats.
Update Readme.md

* Add missing dependencies

* Fix: Add source to file header

Parsing csv data and json to a software that wants to extract specific
features out of the data file, it is important to know from which
datasource this file came and which version of file format from this
datasource.
Therefor the header is now containing a version as well as a source
type.
To fit excel standards these header informations are marked with # for
csv files.
v1.0.0 documents the old format before this change