Add SaveDeque Class, optimize TransformData, fix unbounded memory growth (#58)
* Add SaveDeque-Class
Type:Save Revised

Add SaveDeque to fix the RAM-Bug
Type:Save wirte now the Data from SaveDeque

* SaveDeque Size can now be adjusted
SaveDeque can now be cleared

new cmd Command -> buffer

* Transformdata optimized
New CLI to set buffer-size

- Replaces two loops with one
- Remove CMD Buffer
- Buffer-size can now be set on the server side with -b <int>

* Add Comments
Add constexpr for Max and Min Size Buffer

* Fixed incorrect line in “saveMeasurement”

* write_csv and write_json edited

Both functions were edited for the CLI command to write data directly to a file.
An IdleCount was implemented to wait for new data and prevent a continuous loop.

* Remove clear() function

The function has been removed because it is not needed.
It caused problems with malloc_trim, which is not supported on Windows.

* IDLE_LIMIT increased
Write messages are now only sent with the verbose flag

IDLE_LIMIT to avoid problems with various devices

* Revise IDLE_LIMIT

The IDLE_LIMIT value now has the correct value.