Recording option implemented
The client now has the option to start a recording with {“type”:“record”,‘set’:“start”}.
If it doesn't already exist, a new folder called Record is created.
A <timestamp>.tmp file is created in the folder, into which the data is written.
The data is pulled from the SaveDeque buffer.
With {“type”:“record”,‘set’:“stop”}, the recording stops and the file is given the right extension.
Help functions for creating the files have been implemented.