Add Record (#59)
* Recording option implemented

The client now has the option to start a recording with {“type”:“record”,‘set’:“start”}.
If it doesn't already exist, a new folder called Record is created.
A <timestamp>.tmp file is created in the folder, into which the data is written.
The data is pulled from the SaveDeque buffer.
With {“type”:“record”,‘set’:“stop”}, the recording stops and the file is given the right extension.
Help functions for creating the files have been implemented.

* Add Comments

* Add missing variable RECORDING

Delete unnecessary if-query
Update communication submodule from latest master version

* Delete unnecessary if-query

* REST API endpoint for download adjusted

Unnecessary function startOneShotSave(m) removed.

<timestamp>.(csv/json) can now be downloaded via the download endpoint. After stopping the recording, the following message is now generated:

{“type”:“file-ready”,‘url’:“/download/Record/<timestamp>.(csv/json)”}

The correct name of the file is then specified.

* Delete unnecessary lines